Move clearWaitForDevice to deviceLister
Because we usually do not put thunks in slices.
