+++ src/index.ts
@@ -96,12 +96,12 @@ export {
 } from './Device/deviceSlice';
 export {
     setWaitForDevice,
-    clearWaitForDevice,
     getAutoReselect,
     getWaitingForDeviceTimeout,
     getWaitingToAutoReselect,
     getWaitForDevice,
 } from './Device/deviceAutoSelectSlice';
+export { clearWaitForDevice } from './Device/deviceLister';
 export { deviceInfo } from './Device/deviceInfo/deviceInfo';
 export { getDeviceLibContext } from './Device/deviceLibWrapper';
 export { isDeviceInDFUBootloader } from './Device/sdfuOperations';
